+++ CHANGELOG.md
@@ -17,7 +17,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 - Updates the `Table` component to add the `columnStyles` prop to allow for custom column styles (i.e. width, text alignment, etc.).
 - Updates the `Table` component to add the `tableTextSize` prop to set the size of the text within the table.
 - Updates the `Table` component to add the `titleText` and `showTitleText` props to control the `<caption>` element and `aria-label` attribute.
-- Updates the styles for the `Template` component to use `display` and `overflow` for the primary content area to better accommodate the horizontal scrolling in the `Table` component.
+- Updates the styles for the `Template` component to better accommodate the horizontal scrolling in the `Table` component.
 
 ## 3.3.2 (September 19, 2024)